Add Intake and improve vision

+++ src/main/java/frc/robot/subsystems/vision/VisionConstants.java
@@ -11,7 +11,7 @@ import java.util.Map;
 
 public class VisionConstants {
   /** Map of Camera IDs (from C) to Human Readable Names */
-  public static final Map<Long, String> CAMERA_MAP = Map.of(1L, "TestCamera");
+  public static final Map<Integer, String> CAMERA_MAP = Map.of(1, "TestCamera");
 
   /** Time in seconds before we consider a camera offline */
   public static final double OFFLINE_TIMEOUT_SECONDS = 1.0;
